pyproject.toml
[project]
name = "youtubefilter"
version = "0.1.0"
description = ""
authors = [
    {name = "Henning Agge",email = "yagni@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "blinker (>=1.9.0,<2.0.0)",
    "google-api-python-client (>=2.188.0,<3.0.0)",
    "google-api-core (>=2.29.0,<3.0.0)",
    "google-auth (>=2.48.0,<3.0.0)",
    "google-auth-httplib2 (>=0.3.0,<0.4.0)",
    "google-auth-oauthlib (>=1.2.4,<2.0.0)",
    "googleapis-common-protos (>=1.72.0,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "certifi (>=2026.1.4,<2027.0.0)",
    "charset-normalizer (>=3.4.4,<4.0.0)",
    "click (>=8.3.1,<9.0.0)",
    "httplib2 (>=0.31.2,<0.32.0)",
    "idna (>=3.11,<4.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "markupsafe (>=3.0.3,<4.0.0)",
    "oauthlib (>=3.3.1,<4.0.0)",
    "proto-plus (>=1.27.0,<2.0.0)",
    "protobuf (>=6.33.5,<7.0.0)",
    "pyasn1 (>=0.6.2,<0.7.0)",
    "pyasn1-modules (>=0.4.2,<0.5.0)",
    "pyparsing (>=3.3.2,<4.0.0)",
    "requests-oauthlib (>=2.0.0,<3.0.0)",
    "uritemplate (>=4.2.0,<5.0.0)",
    "urllib3 (>=2.6.3,<3.0.0)",
    "werkzeug (>=3.1.5,<4.0.0)",
    "google-genai (>=1.62.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "bandit (>=1.9.3,<2.0.0)",
    "safety (>=3.7.0,<4.0.0)",
    "detect-secrets (>=1.5.0,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "autopep8 (>=2.3.2,<3.0.0)"
]
